+++ core/src/main/kotlin/org/sert2521/sertain/coroutines/RobotScope.kt
@@ -3,8 +3,6 @@ package org.sert2521.sertain.coroutines
 import kotlinx.coroutines.CoroutineScope
 import kotlin.coroutines.CoroutineContext
 
-open class RobotScope : CoroutineScope {
+object RobotScope : CoroutineScope {
     override val coroutineContext: CoroutineContext = RobotDispatcher
-
-    companion object : RobotScope()
 }
