+++ core/src/main/kotlin/org/sert2521/sertain/subsystems/Subsystem.kt
@@ -8,23 +8,13 @@ import org.sert2521.sertain.events.Use
 import org.sert2521.sertain.events.fire
 import kotlin.coroutines.coroutineContext
 
-abstract class Subsystem(val name: String) {
+abstract class Subsystem(val name: String, val default: (suspend () -> Unit)? = null) {
     var isEnabled = true
 
     internal var currentJob: Job? = null
 
     val occupied: Boolean
         get() = currentJob != null
-
-    internal var default: (suspend CoroutineScope.() -> Unit)? = null
-
-    fun default(action: suspend CoroutineScope.() -> Unit) {
-        default = action
-    }
-
-    fun noDefault() {
-        default = null
-    }
 }
 
 suspend fun <R> use(
