+++ core/src/main/kotlin/org/sert2521/sertain/subsystems/SubsystemManager.kt
@@ -3,6 +3,7 @@ package org.sert2521.sertain.subsystems
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.CoroutineStart
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.NonCancellable
 import kotlinx.coroutines.coroutineScope
@@ -17,7 +18,10 @@ import kotlin.coroutines.AbstractCoroutineContextElement
 import kotlin.coroutines.CoroutineContext
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
+import kotlin.reflect.KClass
+import kotlin.reflect.full.createInstance
 
+@ExperimentalCoroutinesApi
 fun CoroutineScope.manageSubsystems() {
     subscribe<Use<Any?>> { use ->
         // Subsystems used in parent coroutine
@@ -82,7 +86,7 @@ fun CoroutineScope.manageSubsystems() {
                     it.currentJob = null
                     if (it.default != null) {
                         RobotScope.launch {
-                            use(it, name = "DEFAULT") { it.default?.invoke(this) }
+                            use(it, name = "DEFAULT") { it.default.invoke() }
                         }
                     }
                 }
@@ -93,4 +97,35 @@ private class Requirements(
     requirements: Set<Subsystem>
 ) : Set<Subsystem> by requirements, AbstractCoroutineContextElement(Key) {
     companion object Key : CoroutineContext.Key<Requirements>
+}
+
+internal var subsystems = mutableMapOf<KClass<*>, Subsystem>()
+
+fun <S : Subsystem> add(reference: KClass<*>, subsystem: S) {
+    subsystems[reference] = subsystem
+}
+
+fun <S : Subsystem> add(subsystem: S) {
+    add(subsystem::class, subsystem)
+}
+
+inline fun <reified S : Subsystem> add() {
+    add(S::class, S::class.createInstance())
+}
+
+fun <S : Subsystem> access(reference: KClass<S>): S {
+    @Suppress("unchecked_cast") // Safe because subsystems is internally managed
+    subsystems[reference]?.let {
+        return it as S
+    }
+    throw IllegalStateException("Subsystem with type ${reference.qualifiedName} does not exist." +
+            " Did you forget to add it?")
+}
+
+inline fun <reified S : Subsystem> access() = access(S::class)
+
+inline fun <reified S : Subsystem> TaskConfigure.use(): S {
+    val subsystem = access(S::class)
+    this += subsystem
+    return subsystem
 }
