+++ core/src/main/kotlin/org/sert2521/sertain/subsystems/Tasks.kt
@@ -25,7 +25,7 @@ suspend fun doTask(name: String = "ANONYMOUS_TASK", configure: TaskConfigure.() 
     }
 }
 
-fun <S : Subsystem> TaskConfigure.use(subsystem: S): S {
-    this += subsystem
-    return subsystem
-}
+// fun <S : Subsystem> TaskConfigure.use(subsystem: S): S {
+//    this += subsystem
+//    return subsystem
+// }
